Code refactoring: move cypress shell types to cypress/utils/shell
