--- cypress/support/types/shell.ts
@@ -1,4 +1,0 @@
-/**
- * Mirrors shell/utils/pagination-utils.ts
- */
-export const PAGINATION_UTILS = { defaultPageSize: 100000 };
